[EI-201] [api] refactor: 폴더 구조 변경
