+++ api/src/numerical-guidance/application/query/indicator/get-indicator-list/dto/etf.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IndicatorType } from '../../../../../utils/type/type-definition';
+import { IndicatorType } from '../../../../../../utils/type/type-definition';
 
-export class FundDto {
+export class ETFDto {
   @ApiProperty({
     example: 'c6a99067-27d0-4358-b3d5-e63a64b604c0',
     description: '지표 id',
@@ -15,19 +15,19 @@ export class FundDto {
   index: number;
 
   @ApiProperty({
-    example: 'funds',
+    example: 'etf',
     description: '지표 타입',
   })
   indicatorType: IndicatorType;
 
   @ApiProperty({
-    example: '0P00000AMG',
+    example: 'SPY',
     description: '지표 심볼',
   })
   symbol: string;
 
   @ApiProperty({
-    example: 'Morgan Stanley Investment Funds - Sustainable Asia Equity Fund B',
+    example: 'SPDR S&P 500 ETF Trust',
     description: '지표명',
   })
   name: string;
@@ -45,16 +45,16 @@ export class FundDto {
   currency: string;
 
   @ApiProperty({
-    example: 'OTC',
+    example: 'NYSE',
     description: '거래소 종류',
   })
   exchange: string;
 
   @ApiProperty({
-    example: 'Mutual Fund',
-    description: '채권 타입',
+    example: 'XNYS',
+    description: '증권 거래소 식별 코드',
   })
-  type: string;
+  mic_code: string;
 
   constructor(
     id: string,
@@ -65,7 +65,7 @@ export class FundDto {
     country: string,
     currency: string,
     exchange: string,
-    type: string,
+    mic_code: string,
   ) {
     this.id = id;
     this.index = index;
@@ -75,10 +75,10 @@ export class FundDto {
     this.country = country;
     this.currency = currency;
     this.exchange = exchange;
-    this.type = type;
+    this.mic_code = mic_code;
   }
 
-  public static create({ id, index, indicatorType, symbol, name, country, currency, exchange, type }) {
-    return new FundDto(id, index, indicatorType, symbol, name, country, currency, exchange, type);
+  public static create({ id, index, indicatorType, symbol, name, country, currency, exchange, mic_code }) {
+    return new ETFDto(id, index, indicatorType, symbol, name, country, currency, exchange, mic_code);
   }
 }
