+++ api/src/numerical-guidance/application/query/indicator/get-indicator-list/get-indicator-list.query.ts
@@ -1,5 +1,5 @@
 import { IQuery } from '@nestjs/cqrs';
-import { IndicatorType } from '../../../../utils/type/type-definition';
+import { IndicatorType } from '../../../../../utils/type/type-definition';
 
 export class GetIndicatorListQuery implements IQuery {
   constructor(
