+++ api/src/numerical-guidance/infrastructure/adapter/persistence/indicator/indicator.persistent.adapter.ts
@@ -24,15 +24,15 @@ import { ForexPairEntity } from './entity/forex-pair.entity';
 import { FundEntity } from './entity/fund.entity';
 import { IndicesEntity } from './entity/indices.entity';
 import { StockEntity } from './entity/stock.entity';
-import { BondsDto } from '../../../../application/query/indicator/dto/bonds.dto';
+import { BondsDto } from '../../../../application/query/indicator/get-indicator-list/dto/bonds.dto';
 import { CursorPageMetaDto } from '../../../../../utils/pagination/cursor-page.meta.dto';
 import { CursorPageDto } from '../../../../../utils/pagination/cursor-page.dto';
-import { CryptoCurrenciesDto } from '../../../../application/query/indicator/dto/crypto-currencies.dto';
-import { ETFDto } from '../../../../application/query/indicator/dto/etf.dto';
-import { ForexPairDto } from '../../../../application/query/indicator/dto/forex-pair.dto';
-import { IndicesDto } from '../../../../application/query/indicator/dto/indices.dto';
-import { StockDto } from '../../../../application/query/indicator/dto/stock.dto';
-import { FundDto } from '../../../../application/query/indicator/dto/fund.dto';
+import { CryptoCurrenciesDto } from '../../../../application/query/indicator/get-indicator-list/dto/crypto-currencies.dto';
+import { ETFDto } from '../../../../application/query/indicator/get-indicator-list/dto/etf.dto';
+import { ForexPairDto } from '../../../../application/query/indicator/get-indicator-list/dto/forex-pair.dto';
+import { IndicesDto } from '../../../../application/query/indicator/get-indicator-list/dto/indices.dto';
+import { StockDto } from '../../../../application/query/indicator/get-indicator-list/dto/stock.dto';
+import { FundDto } from '../../../../application/query/indicator/get-indicator-list/dto/fund.dto';
 import { CryptoCurrenciesMapper } from './mapper/crypto-currencies.mapper';
 import { ForexPairMapper } from './mapper/forex-pair.mapper';
 import { IndicesMapper } from './mapper/indices.mapper';
